Implement Bricks PlayerPrefs
PlayerPrefs is a class that comes with unity that acts as a global dictionary able to store values and be accessed elsewhere. Useful for settings. We did an example of this making an options scene and a volume slider.

Known bug: The slider will remain where I last had it in-between runs (it wont restart at default)