Begin the Main Menu and First Level
I have begun the first level as well as making the main menu. I added scripts and a cinemachine camera for the first level scene. Struggling with a mechanic to have the Safty Platform rise when the player approaches. No current sprites for the player so its just a capsule for now.