Fix Safety Platform and Jumping
I have fixed the Safety Platform so now it will remain where it is when the player is a certain distance away from it. I have also fixed jumping so that it applies a force vector to the player.

Possible Issue: The player can double jump. For now I think its a nifty feature so I will keep it but I'm not sure why its doing it.