Add Level 1 layout
Added a forest pixel sprites for a level 1 layout. Fixed some bugs in regards to the safety platform so that it works in the new level.

Double jump issue was removed by adding OnCollisionExit2D and setting the boolean to false when the sprite left the collison.

Current Issues:
- You can wall climb by repeatedly press jump.
- Need character sprite and animation
- Platform detect what way player is facing

Future additions:
- Sign/character telling what the player can and cant do in controls.